requirements.txt
gitdb2==2.0.4
GitPython==2.1.11
pyperclip==1.6.4
smmap2==2.0.4
